lib: Fix find_test()
The find_test function did not terminate if no match
was found due to use of unsigned type for positioning.

Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16456

+++ lib/ktf_int.cpp
@@ -281,7 +281,7 @@ KernelTest* KernelTestMgr::find_test(const std::string&setname,
 				     const std::string& testname,
 				     std::string* pctx)
 {
-  size_t pos;
+  off_t pos;
   log(KTF_DEBUG, "find test %s.%s\n", setname.c_str(), testname.c_str());
 
   /* Try direct lookup first: */
